LunaCommon/requirements.txt
ahk==0.11.1
beautifulsoup4==4.9.3
certifi==2020.6.20
cffi==1.15.1
chardet==3.0.4
click==7.1.2
comtypes==1.1.7
docopt==0.6.2
future==0.18.2
idna==2.10
Jinja2==2.11.2
joblib==0.17.0
Js2Py==0.70
MarkupSafe==1.1.1
MouseInfo==0.1.3
mss==6.0.0
neo4j-driver==4.3.6
neobolt==1.7.17
neomodel==4.0.8
nltk==3.5
numpy==1.19.2
opencv-python==4.4.0.44
packaging==20.4
Pillow==7.2.0
pip==20.2.3
pipwin==0.5.0
PyAudio==0.2.11
PyAutoGUI==0.9.51
pycparser==2.21
pydub==0.25.1
PyGetWindow==0.0.9
pyglet==1.3.3
pyjsparser==2.7.1
PyMsgBox==1.0.8
pyparsing==2.4.7
pyperclip==1.8.0
pypiwin32==223
PyPrind==2.11.2
PyRect==0.1.4
PyScreeze==0.1.26
pySmartDL==1.3.4
pyttsx3==2.71
PyTweening==1.0.3
pytz==2022.1
pywin32==228
ratcave==1.0.0
redis==3.5.3
regex==2020.9.27
requests==2.24.0
scipy==1.5.2
setuptools==49.2.1
Shapely==1.7.1
simpleaudio==1.0.4
six==1.16.0
SoundFile==0.10.3.post1
soupsieve==2.0.1
speech-recognition-fork==3.8.1.2021.6.14
tqdm==4.50.0
tzlocal==2.1
urllib3==1.25.10
vosk==0.3.7
wavefront-reader==0.2.2
wheel==0.35.1
